Layers separated from models. Not tested, any issue will show up later.
